Create AssignmentNo2.cpp
Department maintains a student information. The file contains roll number, name, division and address. 
Allow user to add, delete information of student. Display information of particular students.
If record of student does not exist an appropriate message is displayed. 
If it is, then the system displays the student details. Use sequential file to maintain data.